+++ example/__about_importlib__.py
@@ -1,17 +1,16 @@
 """Metadata for toml."""
 
 __all__ = [
-    "__metadata_version__",
-    "__title__",
-    "__version__",
-    "__home_page__",
     "__author__",
     "__author_email__",
-    "__license__",
-    "__platform__",
-    "__classifier__",
-    "__requires_python__",
     "__description__",
+    "__home_page__",
+    "__license__",
+    "__metadata_version__",
+    "__platform__",
+    "__requires_python__",
+    "__title__",
+    "__version__",
 ]
 
 __metadata_version__ = "2.1"
@@ -22,6 +21,5 @@ __author__ = "William Pearson"
 __author_email__ = "uiri@xqz.ca"
 __license__ = "MIT"
 __platform__ = "UNKNOWN"
-__classifier__ = "Programming Language :: Python :: Implementation :: PyPy"
 __requires_python__ = ">=2.6, !=3.0.*, !=3.1.*, !=3.2.*"
 __description__ = "Python Library for Tom's Obvious, Minimal Language"