+++ example/__about_importlib_for_mmm__.py
@@ -1,31 +1,25 @@
 """Metadata for metametameta."""
 
 __all__ = [
-    "__metadata_version__",
-    "__title__",
-    "__version__",
-    "__project_url__",
     "__author_email__",
+    "__description__",
+    "__description_content_type__",
+    "__keywords__",
     "__license__",
     "__license_file__",
-    "__keywords__",
-    "__classifier__",
+    "__metadata_version__",
     "__requires_python__",
-    "__requires_dist__",
-    "__description_content_type__",
-    "__description__",
+    "__title__",
+    "__version__",
 ]
 
 __metadata_version__ = "2.4"
 __title__ = "metametameta"
-__version__ = "0.1.11"
-__project_url__ = "Change Log, https://github.com/matthewdeanmartin/metametameta/blob/main/CHANGELOG.md"
+__version__ = "0.1.13"
 __author_email__ = "Matthew Martin <matthewdeanmartin@gmail.com>"
 __license__ = "MIT"
 __license_file__ = "LICENSE"
 __keywords__ = "metadata,packaging"
-__classifier__ = "Programming Language :: Python :: 3.14"
-__requires_python__ = ">=3.8"
-__requires_dist__ = "totalhelp>=0.1.1"
+__requires_python__ = ">=3.9"
 __description_content_type__ = "text/markdown"
 __description__ = "Generate __about__.py with dunder meta."